requirements.txt
crewai
groq
streamlit
requests
gspread
oauth2client
python-dotenv
smtplib
email
